requirements.txt
dill==0.4.0
names==0.3.0
psutil==7.0.0
pytest==8.4.1
PyYAML==6.0.2
PyYAML==6.0.2
